requirements.txt
# Core
python-dotenv==1.0.0
pytz==2023.3

# Database
mysql-connector-python==8.2.0
SQLAlchemy==2.0.23
pymysql==1.1.0

# Data Processing
pandas==2.1.3
numpy==1.26.2
lxml==4.9.3

# Streamlit & Visualization
streamlit==1.29.0
plotly==5.18.0
altair==5.2.0

# Utilities
colorlog==6.8.0
watchdog==3.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0